# apps/web/app/api/v1/venues/[id]/route.ts
/**
 * GET /api/v1/venues/[id]
 * Get single venue details
 */

import { NextRequest, NextResponse } from 'next/server';
import { prisma } from '@/lib/prisma';

interface VenueDetails {
  id: string;
  name: string;
  address: string;
  city: string;
  state: string;
  zip: string;
  capacity: number;
  amenities: string[];
  contact: {
    phone?: string;
    email?: string;
    website?: string;
  };
  statistics: {
    total_tournaments: number;
    active_tournaments: number;
    total_players_hosted: number;
  };
}

export async function GET(
  request: NextRequest,
  { params }: { params: Promise<{ id: string }> }
) {
  try {
    const { id: venueId } = await params;

    // TODO: Add API key authentication middleware
    // TODO: Add rate limiting
    // TODO: Add tenant scoping from API key

    // For now, use a placeholder tenant ID
    const tenantId = 'placeholder-tenant-id';

    // Note: This assumes venues table exists in schema
    // If not, this endpoint will need to be updated

    // Mock response - replace with actual venue query when schema is updated
    // TODO: Implement actual venue query:
    // const venue = await prisma.venue.findUnique({
    //   where: {
    //     id: venueId,
    //     orgId: tenantId,
    //   },
    //   include: {
    //     tournaments: {
    //       select: {
    //         id: true,
    //         status: true,
    //       },
    //     },
    //   },
    // });

    // if (!venue) {
    //   return NextResponse.json(
    //     {
    //       error: {
    //         code: 'not_found',
    //         message: 'Venue not found',
    //       },
    //     },
    //     { status: 404 }
    //   );
    // }

    return NextResponse.json(
      {
        error: {
          code: 'not_implemented',
          message:
            'Venues table not yet implemented in database schema. This endpoint will be available once the venues table is added.',
        },
      },
      { status: 501 }
    );
  } catch (error) {
    console.error('Venue details API error:', error);
    return NextResponse.json(
      {
        error: {
          code: 'internal_error',
          message: 'An error occurred while fetching venue details',
        },
      },
      { status: 500 }
    );
  }
}